Implemented mod operation for big numbers (no idea whether the implementation works...). Added project reference of JpegDecoder to MiniStdlib.

+++ src/BigNumber/mulUBUC.cpp
@@ -4,7 +4,7 @@
 
 uint32_t mulUBUCKernel(uint32_t in_carry, uint32_t* in_pCurrentDigit, uint32_t in_n)
 {
-	uint64_t prod = ((uint64_t) in_n) * ((uint64_t) *in_pCurrentDigit)+in_carry;
+	uint64_t prod = ((uint64_t) in_n) * ((uint64_t) *in_pCurrentDigit)+(uint64_t) in_carry;
 	*in_pCurrentDigit = (uint32_t) prod;
 
 	return (uint32_t) (prod >> 32);
